[ch32781] Donor - refactor Gavi page access, add MOU multiple selection filter
