+++ src/pages/reports/search-constants.jsx
@@ -46,7 +46,7 @@ export const MODIFIED_AFTER_FIELD = 'modified__gte';
 export const COUNTRY_NAME = 'country_name__contains';
 export const CTN_NUMBER = 'number__contains';
 export const GAVI_WBS = 'g_a_v_i_w_b_s__contains';
-export const MOU_NUMBER = 'm_o_u_number';
+export const MOU_REFERENCE = 'm_o_u_r_eference';
 export const SENT_GAVI_DATE_BEFORE_FIELD = 'sent_to_g_a_v_i_date__lte';
 export const SENT_GAVI_DATE_AFTER_FIELD = 'sent_to_g_a_v_i_date__gte';
 export const PREPAID_STATUS = 'prepaid_status';
